Rename module `backend` to `base`
